refactor: separate id alloc implemenatations from setup.rs
